Additional refactor of widgets
